Fix build, app launching, config persistence, and UI bugs
- Fix build: Update payload.asm to use correct CDL app list (math, usr32,
  syskernel, proc, timer, gui, sysmon, jsengine, netdiag) instead of old
  terminal/files/waterhole/nettools/textedit/browser CDL files
- Fix app launching: Unify execute_program() to use resolve_and_load()
  via wrap_exec() instead of separate duplicate path resolution logic.
  This ensures consistent app resolution from dock, desktop, and shell.
- Fix wrap_cdl_load: Return actual slot index instead of hardcoded 1
  when library is already loaded, fixing symbol resolution bugs
- Fix config persistence: Add delete-before-create for config files,
  verify write by reading back, and double-flush to ensure data reaches
  persistent storage across reboots
- Fix context menu: Correct item_count from 5 to 6 for File/Icon context
  menu so 'Delete' option is visible (was writing to index 5 but only
  rendering 5 items)
- Fix duplicate SNAP_MARGIN definition in bubbleview.c
- Remove old CDL app files (terminal, files, waterhole, nettools,
  textedit, browser) that were superseded by new CDL system libraries

+++ core/tcp.c
@@ -397,7 +397,7 @@ void tcp_conn_set_data_callback(void* conn_ptr, void (*callback)(uint8_t*, uint1
 }
 
 // Set state change callback for a TCP connection
-void tcp_conn_set_state_callback(void* conn_ptr, void (*callback)(int, int)) {
+void tcp_conn_set_state_callback(void* conn_ptr, void (*callback)(uint8_t, uint8_t)) {
     tcp_connection_t* conn = (tcp_connection_t*)conn_ptr;
     if (conn) {
         conn->on_state_change = callback;
